feat: add /api/users test case, remove:removed load balancer and created app.ts
