+++ package.json
@@ -3,28 +3,24 @@
   "version": "1.0.0",
   "main": "index.ts",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
     "build": "webpack",
-    "dev": "ts-node-dev ./src/index.ts"
+    "dev": "ts-node-dev ./src/index.ts",
+    "test": "jest --detectOpenHandles"
+
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "description": "",
   "devDependencies": {
-    "@types/chai": "^5.0.0",
     "@types/dotenv": "^6.1.1",
     "@types/jest": "^29.5.13",
     "@types/supertest": "^6.0.2",
     "@types/uuid": "^10.0.0",
     "@types/webpack": "^5.28.5",
-    "chai": "^5.1.1",
     "cross-env": "^7.0.3",
-    "jest": "^29.7.0",
-    "mocha": "^10.7.3",
     "nodemon": "^3.1.7",
     "supertest": "^7.0.0",
-    "ts-jest": "^29.2.5",
     "ts-node": "^10.9.2",
     "ts-node-dev": "^2.0.0",
     "typescript": "^5.6.3",
